Removed my definitely on purpose Rube Goldberg mask function and storage
Was a bit silly to store masks as uint64_t and then reference their positions with uint8_t index, when that index == the bit pattern at the indexed position... Would be a good way to store non-sequential bit patterns, or patterns that weren't uint8_t, but definitely not a good way to store sequential uint8_t's. Refactored and completed prototype OpNode execution.